[NU-1979] DeploymentService refactor: extracted ActionService and ScenarioStateProvider + BuildInfo as a Map -> ModelInfo as a value class
