+++ designer/restmodel/src/main/scala/pl/touk/nussknacker/restmodel/component/ComponentApiEndpoints.scala
@@ -1,6 +1,5 @@
 package pl.touk.nussknacker.restmodel.component
 
-import pl.touk.nussknacker.engine.api.ProcessVersion
 import pl.touk.nussknacker.engine.api.component.Component.AllowedProcessingModes
 import pl.touk.nussknacker.engine.api.component.{
   ComponentGroupName,
@@ -14,17 +13,18 @@ import pl.touk.nussknacker.engine.api.deployment.{
   ProcessActionState,
   ScenarioActionName
 }
+import pl.touk.nussknacker.engine.api.modelinfo.ModelInfo
 import pl.touk.nussknacker.engine.api.process.{ProcessId, ProcessName, VersionId}
 import pl.touk.nussknacker.restmodel.BaseEndpointDefinitions
 import pl.touk.nussknacker.restmodel.BaseEndpointDefinitions.SecuredEndpoint
 import pl.touk.nussknacker.restmodel.component.NodeUsageData.ScenarioUsageData
 import pl.touk.nussknacker.security.AuthCredentials
 import sttp.model.StatusCode.{NotFound, Ok}
+import sttp.tapir.Codec.PlainCodec
+import sttp.tapir.EndpointIO.Example
 import sttp.tapir._
 import sttp.tapir.generic.auto.schemaForCaseClass
 import sttp.tapir.json.circe.jsonBody
-import sttp.tapir.Codec.PlainCodec
-import sttp.tapir.EndpointIO.Example
 
 import java.net.URI
 import java.time.Instant
@@ -37,6 +37,7 @@ class ComponentApiEndpoints(auth: EndpointInput[AuthCredentials]) extends BaseEn
   implicit val versionIdSchema: Schema[VersionId]                   = Schema.schemaForLong.as[VersionId]
   implicit val actionIdSchema: Schema[ProcessActionId]              = Schema.schemaForUUID.as[ProcessActionId]
   implicit val componentGroupNameSchema: Schema[ComponentGroupName] = Schema.string[ComponentGroupName]
+  implicit val modelInfoSchema: Schema[ModelInfo]                   = Schema.schemaForMap[String].as[ModelInfo]
 
   import ComponentApiEndpoints.ComponentCodec._
 
@@ -121,7 +122,7 @@ class ComponentApiEndpoints(auth: EndpointInput[AuthCredentials]) extends BaseEn
                         failureMessage = None,
                         commentId = None,
                         comment = None,
-                        buildInfo = Map.empty
+                        modelInfo = None
                       )
                     )
                   )
