+++ designer/server/src/main/resources/db/migration/hsql/V1_063__RenameBuildInfoToModelInfo.sql
@@ -1,0 +1,1 @@
+ALTER TABLE "scenario_activities" ALTER COLUMN "build_info" RENAME TO "model_info";
