+++ designer/server/src/main/resources/db/migration/postgres/V1_063__RenameBuildInfoToModelInfo.sql
@@ -1,0 +1,1 @@
+ALTER TABLE "scenario_activities" RENAME COLUMN "build_info" TO "model_info";
