+++ designer/server/src/main/scala/pl/touk/nussknacker/ui/db/entity/BaseEntityFactory.scala
@@ -1,5 +1,7 @@
 package pl.touk.nussknacker.ui.db.entity
 
+import com.typesafe.scalalogging.LazyLogging
+import pl.touk.nussknacker.engine.api.modelinfo.ModelInfo
 import pl.touk.nussknacker.engine.api.deployment.ProcessActionState.ProcessActionState
 import pl.touk.nussknacker.engine.api.deployment.{
   DeploymentStatusName,
@@ -13,7 +15,7 @@ import slick.jdbc.JdbcProfile
 
 import java.util.UUID
 
-trait BaseEntityFactory {
+trait BaseEntityFactory extends LazyLogging {
   protected val profile: JdbcProfile
   import profile.api._
 
@@ -40,5 +42,19 @@ trait BaseEntityFactory {
 
   implicit def deploymentStatusName: BaseColumnType[DeploymentStatusName] =
     MappedColumnType.base[DeploymentStatusName, String](_.value, DeploymentStatusName.apply)
+
+  implicit def modelInfoMapper: BaseColumnType[ModelInfo] =
+    MappedColumnType.base[ModelInfo, String](
+      _.asJsonString,
+      ModelInfo
+        .parseJsonString(_)
+        .fold(
+          { err =>
+            logger.warn(s"Saved model info is not a json's object: ${err.getMessage}. Empty map will be returned")
+            ModelInfo.empty
+          },
+          identity
+        )
+    )
 
 }
