+++ designer/server/src/main/scala/pl/touk/nussknacker/ui/db/entity/ScenarioActivityEntityFactory.scala
@@ -4,6 +4,7 @@ import enumeratum.EnumEntry.UpperSnakecase
 import enumeratum._
 import io.circe.Decoder
 import io.circe.syntax.EncoderOps
+import pl.touk.nussknacker.engine.api.modelinfo.ModelInfo
 import pl.touk.nussknacker.engine.api.deployment.ProcessActionState.ProcessActionState
 import pl.touk.nussknacker.engine.api.deployment._
 import pl.touk.nussknacker.engine.api.process.ProcessId
@@ -59,7 +60,7 @@ trait ScenarioActivityEntityFactory extends BaseEntityFactory {
 
     def errorMessage: Rep[Option[String]] = column[Option[String]]("error_message")
 
-    def buildInfo: Rep[Option[String]] = column[Option[String]]("build_info")
+    def modelInfo: Rep[Option[ModelInfo]] = column[Option[ModelInfo]]("model_info")
 
     def additionalProperties: Rep[AdditionalProperties] = column[AdditionalProperties]("additional_properties")
 
@@ -82,7 +83,7 @@ trait ScenarioActivityEntityFactory extends BaseEntityFactory {
         performedAt,
         state,
         errorMessage,
-        buildInfo,
+        modelInfo,
         additionalProperties,
       ) <> (
         ScenarioActivityEntityData.apply _ tupled, ScenarioActivityEntityData.unapply
@@ -195,6 +196,6 @@ final case class ScenarioActivityEntityData(
     finishedAt: Option[Timestamp],
     state: Option[ProcessActionState],
     errorMessage: Option[String],
-    buildInfo: Option[String],
+    modelInfo: Option[ModelInfo],
     additionalProperties: AdditionalProperties,
 )
