+++ designer/server/src/main/scala/pl/touk/nussknacker/ui/process/ScenarioWithDetailsConversions.scala
@@ -63,7 +63,7 @@ object ScenarioWithDetailsConversions {
         failureMessage = None,
         commentId = None,
         comment = None,
-        buildInfo = Map.empty
+        modelInfo = None
       )
     )
     def getProcessAction(processAction: Option[ProcessAction]) =
