+++ designer/server/src/main/scala/pl/touk/nussknacker/ui/process/repository/ScenarioActionRepositoryAuditLogDecorator.scala
@@ -4,7 +4,8 @@ import db.util.DBIOActionInstances._
 import pl.touk.nussknacker.engine.api.Comment
 import pl.touk.nussknacker.engine.api.deployment.ProcessActionState.ProcessActionState
 import pl.touk.nussknacker.engine.api.deployment._
-import pl.touk.nussknacker.engine.api.process.{ProcessId, ProcessName, ProcessingType, VersionId}
+import pl.touk.nussknacker.engine.api.modelinfo.ModelInfo
+import pl.touk.nussknacker.engine.api.process.{ProcessId, ProcessName, VersionId}
 import pl.touk.nussknacker.ui.process.ScenarioActivityAuditLog
 import pl.touk.nussknacker.ui.security.api.LoggedUser
 import pl.touk.nussknacker.ui.util.FunctorUtils.Ops
@@ -21,10 +22,10 @@ class ScenarioActionRepositoryAuditLogDecorator(underlying: ScenarioActionReposi
       processVersion: VersionId,
       actionName: ScenarioActionName,
       comment: Option[Comment],
-      buildInfoProcessingType: Option[ProcessingType]
+      modelInfo: Option[ModelInfo]
   )(implicit user: LoggedUser): DB[ProcessAction] =
     underlying
-      .addInstantAction(processId, processVersion, actionName, comment, buildInfoProcessingType)
+      .addInstantAction(processId, processVersion, actionName, comment, modelInfo)
       .onSuccessRunAsync(processAction =>
         ScenarioActivityAuditLog.onScenarioImmediateAction(
           processAction.id,
@@ -39,10 +40,10 @@ class ScenarioActionRepositoryAuditLogDecorator(underlying: ScenarioActionReposi
       processId: ProcessId,
       actionName: ScenarioActionName,
       processVersion: Option[VersionId],
-      buildInfoProcessingType: Option[ProcessingType]
+      modelInfo: Option[ModelInfo]
   )(implicit user: LoggedUser): DB[ProcessActionId] =
     underlying
-      .addInProgressAction(processId, actionName, processVersion, buildInfoProcessingType)
+      .addInProgressAction(processId, actionName, processVersion, modelInfo)
       .onSuccessRunAsync(processActionId =>
         ScenarioActivityAuditLog.onScenarioActionStarted(processActionId, processId, actionName, processVersion, user)
       )
@@ -54,7 +55,7 @@ class ScenarioActionRepositoryAuditLogDecorator(underlying: ScenarioActionReposi
       processVersion: VersionId,
       performedAt: Instant,
       comment: Option[Comment],
-      buildInfoProcessingType: Option[ProcessingType]
+      modelInfo: Option[ModelInfo]
   )(implicit user: LoggedUser): DB[Unit] =
     underlying
       .markActionAsFinished(
@@ -64,7 +65,7 @@ class ScenarioActionRepositoryAuditLogDecorator(underlying: ScenarioActionReposi
         processVersion,
         performedAt,
         comment,
-        buildInfoProcessingType
+        modelInfo
       )
       .onSuccessRunAsync(_ =>
         ScenarioActivityAuditLog
@@ -86,7 +87,7 @@ class ScenarioActionRepositoryAuditLogDecorator(underlying: ScenarioActionReposi
       performedAt: Instant,
       comment: Option[Comment],
       failureMessage: String,
-      buildInfoProcessingType: Option[ProcessingType]
+      modelInfo: Option[ModelInfo]
   )(implicit user: LoggedUser): DB[Unit] =
     underlying
       .markActionAsFailed(
@@ -97,7 +98,7 @@ class ScenarioActionRepositoryAuditLogDecorator(underlying: ScenarioActionReposi
         performedAt,
         comment,
         failureMessage,
-        buildInfoProcessingType
+        modelInfo
       )
       .onSuccessRunAsync(_ =>
         ScenarioActivityAuditLog
