+++ designer/server/src/main/scala/pl/touk/nussknacker/ui/process/repository/activities/DbScenarioActivityRepository.scala
@@ -4,6 +4,7 @@ import cats.implicits.catsSyntaxEitherId
 import com.typesafe.scalalogging.LazyLogging
 import db.util.DBIOActionInstances.DB
 import pl.touk.nussknacker.engine.api.Comment
+import pl.touk.nussknacker.engine.api.modelinfo.ModelInfo
 import pl.touk.nussknacker.engine.api.component.ProcessingMode
 import pl.touk.nussknacker.engine.api.deployment.ScenarioAttachment.{AttachmentFilename, AttachmentId}
 import pl.touk.nussknacker.engine.api.deployment._
@@ -474,7 +475,7 @@ class DbScenarioActivityRepository private (override protected val dbRef: DbRef,
       attachmentId: Option[Long] = None,
       comment: Option[String] = None,
       lastModifiedByUserName: Option[String] = None,
-      buildInfo: Option[String] = None,
+      modelInfo: Option[ModelInfo] = None,
       additionalProperties: AdditionalProperties = AdditionalProperties.empty,
   ): ScenarioActivityEntityData = {
     val now = Timestamp.from(clock.instant())
@@ -503,7 +504,7 @@ class DbScenarioActivityRepository private (override protected val dbRef: DbRef,
           }
         case _ => None
       },
-      buildInfo = buildInfo,
+      modelInfo = modelInfo,
       additionalProperties = additionalProperties,
     )
   }
