+++ designer/server/src/test/scala/pl/touk/nussknacker/test/base/it/NuResourcesTest.scala
@@ -66,6 +66,7 @@ import pl.touk.nussknacker.ui.util.{MultipartUtils, NuPathMatchers}
 import slick.dbio.DBIOAction
 
 import java.net.URI
+import java.time.Clock
 import scala.concurrent.{ExecutionContext, Future}
 
 // TODO: Consider using NuItTest with NuScenarioConfigurationHelper instead. This one will be removed in the future.
@@ -115,18 +116,33 @@ trait NuResourcesTest
     futureFetchingScenarioRepository
   )
 
+  protected val scenarioStateProvider: ScenarioStateProvider = ScenarioStateProvider(
+    dmDispatcher,
+    fetchingProcessRepository,
+    actionRepository,
+    dbioRunner,
+    scenarioStateTimeout = None
+  )
+
+  protected val actionService: ActionService = new ActionService(
+    dmDispatcher,
+    fetchingProcessRepository,
+    actionRepository,
+    dbioRunner,
+    processChangeListener,
+    scenarioStateProvider,
+    deploymentCommentSettings,
+    modelInfoProvider,
+    Clock.systemUTC()
+  )
+
   protected val deploymentService: DeploymentService =
     new DeploymentService(
       dmDispatcher,
-      fetchingProcessRepository,
-      actionRepository,
-      dbioRunner,
       processValidatorByProcessingType,
       scenarioResolverByProcessingType,
-      processChangeListener,
-      None,
-      deploymentCommentSettings,
-      mapProcessingTypeDataProvider()
+      actionService,
+      mapProcessingTypeDataProvider(),
     )
 
   protected val processingTypeConfig: ProcessingTypeConfig =
@@ -180,7 +196,7 @@ trait NuResourcesTest
     )
   }
 
-  protected val processService: DBProcessService = createDBProcessService(deploymentService)
+  protected val processService: DBProcessService = createDBProcessService(scenarioStateProvider)
 
   protected val scenarioTestServiceByProcessingType: ProcessingTypeDataProvider[ScenarioTestService, _] =
     mapProcessingTypeDataProvider(
@@ -192,7 +208,7 @@ trait NuResourcesTest
 
   protected val processesRoute = new ProcessesResources(
     processService = processService,
-    processStateService = deploymentService,
+    scenarioStateProvider = scenarioStateProvider,
     processToolbarService = configProcessToolbarService,
     processAuthorizer = processAuthorizer,
     processChangeListener = processChangeListener
@@ -211,7 +227,7 @@ trait NuResourcesTest
     RealLoggedUser(id, name, Map(Category1.stringify -> permissions.toSet))
   }
 
-  protected def createDBProcessService(processStateProvider: ProcessStateProvider): DBProcessService =
+  protected def createDBProcessService(processStateProvider: ScenarioStateProvider): DBProcessService =
     new DBProcessService(
       processStateProvider,
       newProcessPreparerByProcessingType,
@@ -267,7 +283,8 @@ trait NuResourcesTest
   protected def saveCanonicalProcess(process: CanonicalProcess)(
       testCode: => Assertion
   ): Assertion =
-    createProcessRequest(process.name) { _ =>
+    createProcessRequest(process.name) { code =>
+      code shouldBe StatusCodes.Created
       val json = parser.decode[Json](responseAs[String]).rightValue
       val resp = CreateProcessResponse(json)
