+++ designer/server/src/test/scala/pl/touk/nussknacker/test/mock/StubScenarioStateProvider.scala
@@ -1,17 +1,19 @@
 package pl.touk.nussknacker.test.mock
 
 import cats.Traverse
+import db.util.DBIOActionInstances.DB
 import pl.touk.nussknacker.engine.api.deployment._
 import pl.touk.nussknacker.engine.api.process.{ProcessIdWithName, ProcessName, VersionId}
 import pl.touk.nussknacker.restmodel.scenariodetails.ScenarioWithDetails
-import pl.touk.nussknacker.ui.process.ProcessStateProvider
+import pl.touk.nussknacker.ui.process.deployment.ScenarioStateProvider
 import pl.touk.nussknacker.ui.process.repository.ScenarioWithDetailsEntity
 import pl.touk.nussknacker.ui.security.api.LoggedUser
+import slick.dbio.DBIO
 
 import scala.concurrent.Future
 import scala.language.higherKinds
 
-class StubProcessStateProvider(states: Map[ProcessName, ProcessState]) extends ProcessStateProvider {
+class StubScenarioStateProvider(states: Map[ProcessName, ProcessState]) extends ScenarioStateProvider {
 
   override def getProcessState(
       processDetails: ScenarioWithDetailsEntity[_]
@@ -28,5 +30,11 @@ class StubProcessStateProvider(states: Map[ProcessName, ProcessState]) extends P
       implicit user: LoggedUser,
       freshnessPolicy: DataFreshnessPolicy
   ): Future[F[ScenarioWithDetails]] = Future.successful(processTraverse)
+
+  override def getProcessStateDBIO(
+      processDetails: ScenarioWithDetailsEntity[_],
+      currentlyPresentedVersionId: Option[VersionId]
+  )(implicit user: LoggedUser, freshnessPolicy: DataFreshnessPolicy): DB[ProcessState] =
+    DBIO.successful(states(processDetails.name))
 
 }
