+++ designer/server/src/test/scala/pl/touk/nussknacker/ui/notifications/NotificationServiceTest.scala
@@ -3,6 +3,7 @@ package pl.touk.nussknacker.ui.notifications
 import akka.actor.ActorSystem
 import org.mockito.ArgumentMatchers.any
 import org.mockito.Mockito.when
+import org.scalatest.exceptions.TestFailedException
 import org.scalatest.{BeforeAndAfterAll, OptionValues}
 import org.scalatest.funsuite.AnyFunSuite
 import org.scalatest.matchers.should.Matchers
@@ -11,6 +12,7 @@ import pl.touk.nussknacker.engine.api.component.NodesDeploymentData
 import pl.touk.nussknacker.engine.api.deployment.DeploymentUpdateStrategy.StateRestoringStrategy
 import pl.touk.nussknacker.engine.api.deployment._
 import pl.touk.nussknacker.engine.api.deployment.simple.{SimpleProcessStateDefinitionManager, SimpleStateStatus}
+import pl.touk.nussknacker.engine.api.modelinfo.ModelInfo
 import pl.touk.nussknacker.engine.api.process.{ProcessIdWithName, ProcessName, VersionId}
 import pl.touk.nussknacker.engine.build.ScenarioBuilder
 import pl.touk.nussknacker.engine.canonicalgraph.CanonicalProcess
@@ -85,11 +87,7 @@ class NotificationServiceTest
     dbioActionRunner = dbioRunner
   )
 
-  private val actionRepository =
-    DbScenarioActionRepository.create(
-      testDbRef,
-      ProcessingTypeDataProvider.withEmptyCombinedData(Map.empty)
-    )
+  private val actionRepository = DbScenarioActionRepository.create(testDbRef)
 
   private val expectedRefreshAfterSuccess = List(DataToRefresh.activity, DataToRefresh.state)
   private val expectedRefreshAfterFail    = List(DataToRefresh.state)
@@ -104,8 +102,8 @@ class NotificationServiceTest
     val id                = saveSampleProcess(processName)
     val processIdWithName = ProcessIdWithName(id, processName)
 
-    val deploymentManager                        = mock[DeploymentManager]
-    val (deploymentService, notificationService) = createServices(deploymentManager)
+    val deploymentManager                                       = mock[DeploymentManager]
+    val (deploymentService, actionService, notificationService) = createServices(deploymentManager)
 
     def notificationsFor(user: LoggedUser): List[Notification] =
       notificationService
@@ -138,8 +136,11 @@ class NotificationServiceTest
     deployProcess(Success(None), userForSuccess)
     notificationsFor(userForSuccess).map(_.toRefresh) shouldBe List(expectedRefreshAfterSuccess)
 
-    an[Exception] shouldBe thrownBy {
-      deployProcess(Failure(new RuntimeException("Failure")), userForFail)
+    val givenException = new RuntimeException("Failure")
+    intercept[TestFailedException] {
+      deployProcess(Failure(givenException), userForFail)
+    } should matchPattern {
+      case ex: TestFailedException if ex.getCause == givenException =>
     }
 
     notificationsFor(userForFail).map(_.toRefresh) shouldBe List(expectedRefreshAfterFail)
@@ -157,8 +158,8 @@ class NotificationServiceTest
     val id                = saveSampleProcess(processName)
     val processIdWithName = ProcessIdWithName(id, processName)
 
-    val deploymentManager                        = mock[DeploymentManager]
-    val (deploymentService, notificationService) = createServices(deploymentManager)
+    val deploymentManager                                       = mock[DeploymentManager]
+    val (deploymentService, actionService, notificationService) = createServices(deploymentManager)
 
     def notificationsFor(user: LoggedUser): List[Notification] =
       notificationService
@@ -229,8 +230,8 @@ class NotificationServiceTest
     val id                = saveSampleProcess(processName)
     val processIdWithName = ProcessIdWithName(id, processName)
 
-    val deploymentManager                        = mock[DeploymentManager]
-    val (deploymentService, notificationService) = createServices(deploymentManager)
+    val deploymentManager                                       = mock[DeploymentManager]
+    val (deploymentService, actionService, notificationService) = createServices(deploymentManager)
 
     var passedDeploymentId = Option.empty[DeploymentId]
     def deployProcess(
@@ -266,8 +267,8 @@ class NotificationServiceTest
     notificationsAfterDeploy should have length 1
     val deployNotificationId = notificationsAfterDeploy.head.id
 
-    deploymentService
-      .markActionExecutionFinished("Streaming", passedDeploymentId.value.toActionIdOpt.value)
+    actionService
+      .markActionExecutionFinished(Streaming.stringify, passedDeploymentId.value.toActionIdOpt.value)
       .futureValue
     val notificationAfterExecutionFinished =
       notificationService
@@ -310,18 +311,30 @@ class NotificationServiceTest
       config,
       clock
     )
-    val deploymentService = new DeploymentService(
+    val scenarioStateProvider = ScenarioStateProvider(
       managerDispatcher,
       dbProcessRepository,
       actionRepository,
       dbioRunner,
+      scenarioStateTimeout = None
+    )
+    val actionService = new ActionService(
+      managerDispatcher,
+      dbProcessRepository,
+      actionRepository,
+      dbioRunner,
+      mock[ProcessChangeListener],
+      scenarioStateProvider,
+      None,
+      TestFactory.modelInfoProvider,
+      clock
+    )
+    val deploymentService = new DeploymentService(
+      managerDispatcher,
       mock[ProcessingTypeDataProvider[UIProcessValidator, _]],
       mock[ProcessingTypeDataProvider[ScenarioResolver, _]],
-      mock[ProcessChangeListener],
-      None,
-      None,
+      actionService,
       TestFactory.additionalComponentConfigsByProcessingType,
-      clock
     ) {
       override protected def validateBeforeDeploy(
           processDetails: ScenarioWithDetailsEntity[CanonicalProcess],
@@ -350,7 +363,7 @@ class NotificationServiceTest
         )
       }
     }
-    (deploymentService, notificationService)
+    (deploymentService, actionService, notificationService)
   }
 
   private def saveSampleProcess(processName: ProcessName) = {
@@ -363,7 +376,7 @@ class NotificationServiceTest
         processName = processName,
         category = "Default",
         canonicalProcess = sampleScenario,
-        processingType = "Streaming",
+        processingType = Streaming.stringify,
         isFragment = false,
         forwardedUserName = None
       )
