+++ designer/server/src/test/scala/pl/touk/nussknacker/ui/process/deployment/DeploymentServiceSpec.scala
@@ -28,6 +28,7 @@ import pl.touk.nussknacker.test.utils.scalas.DBIOActionValues
 import pl.touk.nussknacker.test.{EitherValuesDetailedMessage, NuScalaTestAssertions, PatientScalaFutures}
 import pl.touk.nussknacker.ui.api.DeploymentCommentSettings
 import pl.touk.nussknacker.ui.listener.ProcessChangeEvent.{OnActionExecutionFinished, OnActionSuccess}
+import pl.touk.nussknacker.ui.process.deployment.ScenarioStateProvider.FragmentStateException
 import pl.touk.nussknacker.ui.process.periodic.flink.FlinkClientStub
 import pl.touk.nussknacker.ui.process.processingtype.ValueWithRestriction
 import pl.touk.nussknacker.ui.process.processingtype.provider.ProcessingTypeDataProvider.noCombinedDataFun
@@ -90,36 +91,50 @@ class DeploymentServiceSpec
 
   private val listener = new TestProcessChangeListener
 
-  private val deploymentCommentSettings = None
+  private val scenarioStateProvider = createScenarioStateProvider(scenarioStateTimeout = None)
 
-  private val deploymentService = createDeploymentService(None)
+  private val actionService = createActionService(deploymentCommentSettings = None)
+
+  private val deploymentService = createDeploymentService()
 
   private val initialVersionId = ProcessVersion.empty.versionId
 
   deploymentManager = MockDeploymentManager.create(
     defaultProcessStateStatus = SimpleStateStatus.Running,
     deployedScenariosProvider = DefaultProcessingTypeDeployedScenariosProvider(testDbRef, "streaming"),
-    actionService = new DefaultProcessingTypeActionService("streaming", deploymentService),
+    actionService = new DefaultProcessingTypeActionService("streaming", actionService),
     scenarioActivityManager = new RepositoryBasedScenarioActivityManager(activityRepository, dbioRunner),
   )
 
   private def createDeploymentService(
-      scenarioStateTimeout: Option[FiniteDuration] = None,
-      deploymentCommentSettings: Option[DeploymentCommentSettings] = deploymentCommentSettings,
+      deploymentCommentSettings: Option[DeploymentCommentSettings] = None,
   ) = {
+    val actionService = createActionService(deploymentCommentSettings)
     new DeploymentService(
+      dmDispatcher,
+      processValidatorByProcessingType,
+      TestFactory.scenarioResolverByProcessingType,
+      actionService,
+      additionalComponentConfigsByProcessingType,
+    )
+  }
+
+  private def createActionService(deploymentCommentSettings: Option[DeploymentCommentSettings]) = {
+    new ActionService(
       dmDispatcher,
       fetchingProcessRepository,
       actionRepository,
       dbioRunner,
-      processValidatorByProcessingType,
-      TestFactory.scenarioResolverByProcessingType,
       listener,
-      scenarioStateTimeout,
+      scenarioStateProvider,
       deploymentCommentSettings,
-      additionalComponentConfigsByProcessingType
+      modelInfoProvider,
+      clock
     )
   }
+
+  private def createScenarioStateProvider(scenarioStateTimeout: Option[FiniteDuration]) =
+    ScenarioStateProvider(dmDispatcher, fetchingProcessRepository, actionRepository, dbioRunner, scenarioStateTimeout)
 
   // TODO: temporary step - we would like to extract the validation and the comment validation tests to external validators
   private def createDeploymentServiceWithCommentSettings = {
@@ -171,7 +186,7 @@ class DeploymentServiceSpec
     )
 
     eventually {
-      val status = deploymentServiceWithCommentSettings
+      val status = scenarioStateProvider
         .getProcessState(processIdWithName, Some(initialVersionId))
         .futureValue
         .status
@@ -202,7 +217,7 @@ class DeploymentServiceSpec
     )
 
     eventually {
-      deploymentServiceWithCommentSettings
+      scenarioStateProvider
         .getProcessState(processIdWithName, Some(initialVersionId))
         .futureValue
         .status shouldBe SimpleStateStatus.Running
@@ -222,7 +237,7 @@ class DeploymentServiceSpec
         .futureValue
 
       eventually {
-        val status = deploymentServiceWithCommentSettings
+        val status = scenarioStateProvider
           .getProcessState(processIdWithName, Some(initialVersionId))
           .futureValue
           .status
@@ -253,14 +268,14 @@ class DeploymentServiceSpec
           )
         )
         .futureValue
-      deploymentService
+      scenarioStateProvider
         .getProcessState(processIdWithName, Some(initialVersionId))
         .futureValue
         .status shouldBe SimpleStateStatus.DuringDeploy
     }
 
     eventually {
-      deploymentService
+      scenarioStateProvider
         .getProcessState(processIdWithName, Some(initialVersionId))
         .futureValue
         .status shouldBe SimpleStateStatus.Running
@@ -274,7 +289,7 @@ class DeploymentServiceSpec
     deploymentManager.withWaitForCancelFinish {
       deploymentService.processCommand(CancelScenarioCommand(CommonCommandData(processId, None, user)))
       eventually {
-        deploymentService
+        scenarioStateProvider
           .getProcessState(processId, Some(initialVersionId))
           .futureValue
           .status shouldBe SimpleStateStatus.DuringCancel
@@ -283,11 +298,10 @@ class DeploymentServiceSpec
   }
 
   test("should mark Action ExecutionFinished and publish an event as finished") {
-
     val processName: ProcessName = generateProcessName
     val (processId, actionId)    = prepareDeployedProcess(processName).dbioActionValues
 
-    deploymentService.markActionExecutionFinished("streaming", actionId).futureValue
+    actionService.markActionExecutionFinished("streaming", actionId).futureValue
     eventually {
       val action =
         actionRepository.getFinishedProcessActions(processId.id, Some(Set(ScenarioActionName.Deploy))).dbioActionValues
@@ -295,7 +309,6 @@ class DeploymentServiceSpec
       action.loneElement.state shouldBe ProcessActionState.ExecutionFinished
       listener.events.toArray.filter(_.isInstanceOf[OnActionExecutionFinished]) should have length 1
     }
-
   }
 
   test("Should mark finished process as finished") {
@@ -303,7 +316,7 @@ class DeploymentServiceSpec
     val (processId, deployActionId) = prepareDeployedProcess(processName).dbioActionValues
 
     checkIsFollowingDeploy(
-      deploymentService.getProcessState(processId, Some(initialVersionId)).futureValue,
+      scenarioStateProvider.getProcessState(processId, Some(initialVersionId)).futureValue,
       expected = true
     )
     fetchingProcessRepository
@@ -316,11 +329,11 @@ class DeploymentServiceSpec
       // we simulate what happens when retrieveStatus is called multiple times to check only one comment is added
       (1 to 5).foreach { _ =>
         checkIsFollowingDeploy(
-          deploymentService.getProcessState(processId, Some(initialVersionId)).futureValue,
+          scenarioStateProvider.getProcessState(processId, Some(initialVersionId)).futureValue,
           expected = false
         )
       }
-      val finishedStatus = deploymentService.getProcessState(processId, Some(initialVersionId)).futureValue
+      val finishedStatus = scenarioStateProvider.getProcessState(processId, Some(initialVersionId)).futureValue
       finishedStatus.status shouldBe SimpleStateStatus.Finished
       finishedStatus.allowedActions shouldBe List(
         ScenarioActionName.Deploy,
@@ -340,12 +353,12 @@ class DeploymentServiceSpec
 
     deploymentManager.withEmptyProcessState(processName) {
       val stateAfterJobRetention =
-        deploymentService.getProcessState(processId, Some(initialVersionId)).futureValue
+        scenarioStateProvider.getProcessState(processId, Some(initialVersionId)).futureValue
       stateAfterJobRetention.status shouldBe SimpleStateStatus.Finished
     }
 
     archiveProcess(processId).dbioActionValues
-    deploymentService
+    scenarioStateProvider
       .getProcessState(processId, Some(initialVersionId))
       .futureValue
       .status shouldBe SimpleStateStatus.Finished
@@ -361,7 +374,7 @@ class DeploymentServiceSpec
         .dbioActionValues
         .flatMap(_.lastStateAction)
         .map(_.actionName) shouldBe expectedAction
-      deploymentService
+      scenarioStateProvider
         .getProcessState(processIdWithName, Some(initialVersionId))
         .futureValue
         .status shouldBe expectedStatus
@@ -446,7 +459,7 @@ class DeploymentServiceSpec
       listener.events shouldBe Symbol("empty")
       // during short period of time, status will be during deploy - because parallelism validation are done in the same critical section as deployment
       eventually {
-        deploymentService
+        scenarioStateProvider
           .getProcessState(processIdWithName, Some(initialVersionId))
           .futureValue
           .status shouldBe SimpleStateStatus.NotDeployed
@@ -459,7 +472,7 @@ class DeploymentServiceSpec
     val (processId, _)           = prepareCanceledProcess(processName).dbioActionValues
 
     deploymentManager.withProcessStateStatus(processName, SimpleStateStatus.Canceled) {
-      deploymentService
+      scenarioStateProvider
         .getProcessState(processId, Some(initialVersionId))
         .futureValue
         .status shouldBe SimpleStateStatus.Canceled
@@ -477,7 +490,7 @@ class DeploymentServiceSpec
       .lastStateAction should not be None
 
     deploymentManager.withEmptyProcessState(processName) {
-      deploymentService
+      scenarioStateProvider
         .getProcessState(processId, Some(initialVersionId))
         .futureValue
         .status shouldBe SimpleStateStatus.Canceled
@@ -502,7 +515,7 @@ class DeploymentServiceSpec
       .lastStateAction should not be None
 
     deploymentManager.withEmptyProcessState(processName) {
-      deploymentService
+      scenarioStateProvider
         .getProcessState(processId, Some(initialVersionId))
         .futureValue
         .status shouldBe SimpleStateStatus.Canceled
@@ -521,7 +534,7 @@ class DeploymentServiceSpec
     val (processId, _)           = prepareCanceledProcess(processName).dbioActionValues
 
     deploymentManager.withProcessStateStatus(processName, SimpleStateStatus.Running) {
-      val state = deploymentService.getProcessState(processId, Some(initialVersionId)).futureValue
+      val state = scenarioStateProvider.getProcessState(processId, Some(initialVersionId)).futureValue
 
       val expectedStatus = ProblemStateStatus.shouldNotBeRunning(true)
       state.status shouldBe expectedStatus
@@ -536,7 +549,7 @@ class DeploymentServiceSpec
     val processId                = prepareProcess(processName).dbioActionValues
 
     deploymentManager.withProcessStateStatus(processName, SimpleStateStatus.Running) {
-      val state = deploymentService.getProcessState(processId, Some(initialVersionId)).futureValue
+      val state = scenarioStateProvider.getProcessState(processId, Some(initialVersionId)).futureValue
       state.status shouldBe SimpleStateStatus.NotDeployed
     }
   }
@@ -546,7 +559,7 @@ class DeploymentServiceSpec
     val (processId, _)           = prepareCanceledProcess(processName).dbioActionValues
 
     deploymentManager.withProcessStateStatus(processName, SimpleStateStatus.DuringCancel) {
-      val state = deploymentService.getProcessState(processId, Some(initialVersionId)).futureValue
+      val state = scenarioStateProvider.getProcessState(processId, Some(initialVersionId)).futureValue
 
       state.status shouldBe SimpleStateStatus.DuringCancel
     }
@@ -560,7 +573,7 @@ class DeploymentServiceSpec
       StatusDetails(SimpleStateStatus.Restarting, None, Some(ExternalDeploymentId("12")), Some(ProcessVersion.empty))
 
     deploymentManager.withProcessStates(processName, List(state)) {
-      val state = deploymentService.getProcessState(processId, Some(initialVersionId)).futureValue
+      val state = scenarioStateProvider.getProcessState(processId, Some(initialVersionId)).futureValue
 
       state.status shouldBe SimpleStateStatus.Restarting
       state.allowedActions shouldBe List(ScenarioActionName.Cancel)
@@ -573,7 +586,7 @@ class DeploymentServiceSpec
     val (processId, _)           = prepareDeployedProcess(processName).dbioActionValues
 
     deploymentManager.withProcessStateStatus(processName, SimpleStateStatus.Canceled) {
-      val state = deploymentService.getProcessState(processId, Some(initialVersionId)).futureValue
+      val state = scenarioStateProvider.getProcessState(processId, Some(initialVersionId)).futureValue
 
       val expectedStatus = ProblemStateStatus.shouldBeRunning(VersionId(1L), "admin")
       state.status shouldBe expectedStatus
@@ -588,7 +601,7 @@ class DeploymentServiceSpec
     val (processId, _)           = prepareDeployedProcess(processName).dbioActionValues
 
     deploymentManager.withEmptyProcessState(processName) {
-      val state = deploymentService.getProcessState(processId, Some(initialVersionId)).futureValue
+      val state = scenarioStateProvider.getProcessState(processId, Some(initialVersionId)).futureValue
 
       val expectedStatus = ProblemStateStatus.shouldBeRunning(VersionId(1L), "admin")
       state.status shouldBe expectedStatus
@@ -613,7 +626,7 @@ class DeploymentServiceSpec
     )
 
     deploymentManager.withProcessStateVersion(processName, SimpleStateStatus.Running, version) {
-      val state = deploymentService.getProcessState(processId, Some(initialVersionId)).futureValue
+      val state = scenarioStateProvider.getProcessState(processId, Some(initialVersionId)).futureValue
 
       val expectedStatus = ProblemStateStatus.mismatchDeployedVersion(VersionId(2L), VersionId(1L), "admin")
       state.status shouldBe expectedStatus
@@ -639,7 +652,7 @@ class DeploymentServiceSpec
 
     // FIXME: doesnt check recover from failed verifications ???
     deploymentManager.withProcessStateVersion(processName, ProblemStateStatus.Failed, version) {
-      val state = deploymentService.getProcessState(processId, Some(initialVersionId)).futureValue
+      val state = scenarioStateProvider.getProcessState(processId, Some(initialVersionId)).futureValue
 
       state.status shouldBe ProblemStateStatus.Failed
       state.allowedActions shouldBe List(ScenarioActionName.Deploy, ScenarioActionName.Cancel)
@@ -651,7 +664,7 @@ class DeploymentServiceSpec
     val (processId, _)           = prepareDeployedProcess(processName).dbioActionValues
 
     deploymentManager.withProcessStateVersion(processName, SimpleStateStatus.Running, Option.empty) {
-      val state = deploymentService.getProcessState(processId, Some(initialVersionId)).futureValue
+      val state = scenarioStateProvider.getProcessState(processId, Some(initialVersionId)).futureValue
 
       val expectedStatus = ProblemStateStatus.missingDeployedVersion(VersionId(1L), "admin")
       state.status shouldBe expectedStatus
@@ -667,7 +680,7 @@ class DeploymentServiceSpec
 
     // FIXME: doesnt check recover from failed future of findJobStatus ???
     deploymentManager.withProcessStateVersion(processName, ProblemStateStatus.FailedToGet, Option.empty) {
-      val state = deploymentService.getProcessState(processId, Some(initialVersionId)).futureValue
+      val state = scenarioStateProvider.getProcessState(processId, Some(initialVersionId)).futureValue
 
       val expectedStatus = ProblemStateStatus.FailedToGet
       state.status shouldBe expectedStatus
@@ -687,7 +700,7 @@ class DeploymentServiceSpec
       .lastStateAction shouldBe None
 
     deploymentManager.withEmptyProcessState(processName) {
-      deploymentService
+      scenarioStateProvider
         .getProcessState(ProcessIdWithName(processId.id, processName), Some(initialVersionId))
         .futureValue
         .status shouldBe SimpleStateStatus.NotDeployed
@@ -709,7 +722,7 @@ class DeploymentServiceSpec
       .lastStateAction shouldBe None
 
     deploymentManager.withEmptyProcessState(processName) {
-      deploymentService
+      scenarioStateProvider
         .getProcessState(processId, Some(initialVersionId))
         .futureValue
         .status shouldBe SimpleStateStatus.NotDeployed
@@ -731,7 +744,7 @@ class DeploymentServiceSpec
       .lastStateAction shouldBe None
 
     deploymentManager.withProcessStateStatus(processName, SimpleStateStatus.Running) {
-      deploymentService
+      scenarioStateProvider
         .getProcessState(ProcessIdWithName(processId.id, processName), Some(initialVersionId))
         .futureValue
         .status shouldBe SimpleStateStatus.NotDeployed
@@ -747,7 +760,7 @@ class DeploymentServiceSpec
     val processName: ProcessName = generateProcessName
     val (processId, _)           = prepareArchivedProcess(processName, None).dbioActionValues
 
-    val state = deploymentService.getProcessState(processId, Some(initialVersionId)).futureValue
+    val state = scenarioStateProvider.getProcessState(processId, Some(initialVersionId)).futureValue
     state.status shouldBe SimpleStateStatus.NotDeployed
   }
 
@@ -758,7 +771,7 @@ class DeploymentServiceSpec
     val (processId, _)           = prepareArchivedProcess(processName, None).dbioActionValues
 
     deploymentManager.withProcessStateStatus(processName, SimpleStateStatus.Running) {
-      val state = deploymentService.getProcessState(processId, Some(initialVersionId)).futureValue
+      val state = scenarioStateProvider.getProcessState(processId, Some(initialVersionId)).futureValue
       state.status shouldBe SimpleStateStatus.NotDeployed
     }
   }
@@ -767,7 +780,7 @@ class DeploymentServiceSpec
     val processName: ProcessName = generateProcessName
     val (processId, _)           = prepareArchivedProcess(processName, Some(Cancel)).dbioActionValues
 
-    val state = deploymentService.getProcessState(processId, Some(initialVersionId)).futureValue
+    val state = scenarioStateProvider.getProcessState(processId, Some(initialVersionId)).futureValue
     state.status shouldBe SimpleStateStatus.Canceled
   }
 
@@ -776,7 +789,7 @@ class DeploymentServiceSpec
     val (processId, _)           = prepareArchivedProcess(processName, Some(Cancel)).dbioActionValues
 
     deploymentManager.withProcessStateStatus(processName, SimpleStateStatus.Running) {
-      val state = deploymentService.getProcessState(processId, Some(initialVersionId)).futureValue
+      val state = scenarioStateProvider.getProcessState(processId, Some(initialVersionId)).futureValue
       state.status shouldBe SimpleStateStatus.Canceled
     }
   }
@@ -785,7 +798,7 @@ class DeploymentServiceSpec
     val processName: ProcessName = generateProcessName
     val (processId, _)           = preparedUnArchivedProcess(processName, None).dbioActionValues
 
-    val state = deploymentService.getProcessState(processId, Some(initialVersionId)).futureValue
+    val state = scenarioStateProvider.getProcessState(processId, Some(initialVersionId)).futureValue
     state.status shouldBe SimpleStateStatus.NotDeployed
   }
 
@@ -796,7 +809,7 @@ class DeploymentServiceSpec
       .addInProgressAction(processId.id, ScenarioActionName.Deploy, Some(VersionId(1)), None)
       .dbioActionValues
 
-    val state = deploymentService.getProcessState(processId, Some(initialVersionId)).futureValue
+    val state = scenarioStateProvider.getProcessState(processId, Some(initialVersionId)).futureValue
     state.status shouldBe SimpleStateStatus.DuringDeploy
   }
 
@@ -807,7 +820,7 @@ class DeploymentServiceSpec
       .fetchLatestProcessesDetails[Unit](ScenarioQuery.empty)
       .dbioActionValues
 
-    val processesDetailsWithState = deploymentService
+    val processesDetailsWithState = scenarioStateProvider
       .enrichDetailsWithProcessState(
         processesDetails
           .map(
@@ -828,7 +841,7 @@ class DeploymentServiceSpec
 
     val statesBasedOnNotCachedInProgressActionTypes =
       processesDetails
-        .map(pd => Option(pd).filterNot(_.isFragment).map(deploymentService.getProcessState).sequence)
+        .map(pd => Option(pd).filterNot(_.isFragment).map(scenarioStateProvider.getProcessState).sequence)
         .sequence
         .futureValue
 
@@ -842,7 +855,7 @@ class DeploymentServiceSpec
     val (processId, _)           = prepareArchivedProcess(processName, None).dbioActionValues
 
     deploymentManager.withProcessStateStatus(processName, SimpleStateStatus.Running) {
-      val state = deploymentService.getProcessState(processId, Some(initialVersionId)).futureValue
+      val state = scenarioStateProvider.getProcessState(processId, Some(initialVersionId)).futureValue
       state.status shouldBe SimpleStateStatus.NotDeployed
     }
   }
@@ -851,7 +864,7 @@ class DeploymentServiceSpec
     val processName: ProcessName = generateProcessName
     val (processId, _)           = prepareArchivedProcess(processName, Some(Deploy)).dbioActionValues
 
-    val state = deploymentService.getProcessState(processId, Some(initialVersionId)).futureValue
+    val state = scenarioStateProvider.getProcessState(processId, Some(initialVersionId)).futureValue
     state.status shouldBe ProblemStateStatus.ArchivedShouldBeCanceled
   }
 
@@ -860,7 +873,7 @@ class DeploymentServiceSpec
     val (processId, _)           = prepareArchivedProcess(processName, Some(Cancel)).dbioActionValues
 
     deploymentManager.withEmptyProcessState(processName) {
-      val state = deploymentService.getProcessState(processId, Some(initialVersionId)).futureValue
+      val state = scenarioStateProvider.getProcessState(processId, Some(initialVersionId)).futureValue
       state.status shouldBe SimpleStateStatus.Canceled
     }
   }
@@ -870,7 +883,7 @@ class DeploymentServiceSpec
     val (processId, _)           = preparedUnArchivedProcess(processName, Some(Cancel)).dbioActionValues
 
     deploymentManager.withProcessStateStatus(processName, SimpleStateStatus.Running) {
-      val state          = deploymentService.getProcessState(processId, Some(initialVersionId)).futureValue
+      val state          = scenarioStateProvider.getProcessState(processId, Some(initialVersionId)).futureValue
       val expectedStatus = ProblemStateStatus.shouldNotBeRunning(true)
       state.status shouldBe expectedStatus
       state.icon shouldBe ProblemStateStatus.icon
@@ -885,7 +898,7 @@ class DeploymentServiceSpec
 
     deploymentManager.withEmptyProcessState(processName) {
       val initialStatus = SimpleStateStatus.NotDeployed
-      deploymentService
+      scenarioStateProvider
         .getProcessState(processIdWithName, Some(initialVersionId))
         .futureValue
         .status shouldBe initialStatus
@@ -899,13 +912,13 @@ class DeploymentServiceSpec
             )
           )
           .futureValue
-        deploymentService
+        scenarioStateProvider
           .getProcessState(processIdWithName, Some(initialVersionId))
           .futureValue
           .status shouldBe SimpleStateStatus.DuringDeploy
 
-        deploymentService.invalidateInProgressActions()
-        deploymentService
+        actionService.invalidateInProgressActions()
+        scenarioStateProvider
           .getProcessState(processIdWithName, Some(initialVersionId))
           .futureValue
           .status shouldBe initialStatus
@@ -918,7 +931,7 @@ class DeploymentServiceSpec
     val (processId, _)           = prepareDeployedProcess(processName).dbioActionValues
 
     val timeout            = 1.second
-    val serviceWithTimeout = createDeploymentService(Some(timeout))
+    val serviceWithTimeout = createScenarioStateProvider(Some(timeout))
 
     val durationLongerThanTimeout = timeout.plus(patienceConfig.timeout)
     deploymentManager.withDelayBeforeStateReturn(durationLongerThanTimeout) {
@@ -934,8 +947,8 @@ class DeploymentServiceSpec
     val processName: ProcessName = generateProcessName
     val id                       = prepareFragment(processName).dbioActionValues
 
-    assertThrowsWithParent[FragmentStateException] {
-      deploymentService.getProcessState(id, Some(initialVersionId)).futureValue
+    assertThrowsWithParent[FragmentStateException.type] {
+      scenarioStateProvider.getProcessState(id, Some(initialVersionId)).futureValue
     }
   }
