+++ docs-internal/api/nu-designer-openapi.yaml
@@ -55,7 +55,6 @@ paths:
                       performedAt: '2023-11-29T08:54:22.520866Z'
                       actionName: DEPLOY
                       state: FINISHED
-                      buildInfo: {}
                 Example1:
                   summary: List component usages with no last Action
                   value:
@@ -5364,7 +5363,7 @@ components:
         version:
           type: string
         processingType:
-          $ref: '#/components/schemas/Map_Map_String_String'
+          $ref: '#/components/schemas/Map_ModelInfo'
         globalBuildInfo:
           anyOf:
           - $ref: '#/components/schemas/Map_String'
@@ -5991,16 +5990,16 @@ components:
         type: array
         items:
           type: string
-    Map_Map_String_String:
-      title: Map_Map_String_String
-      type: object
-      additionalProperties:
-        $ref: '#/components/schemas/Map_String'
     Map_Map_String_TypingResultInJson:
       title: Map_Map_String_TypingResultInJson
       type: object
       additionalProperties:
         $ref: '#/components/schemas/Map_TypingResultInJson'
+    Map_ModelInfo:
+      title: Map_ModelInfo
+      type: object
+      additionalProperties:
+        $ref: '#/components/schemas/Map_String'
     Map_NodeId_String:
       title: Map_NodeId_String
       type: object
@@ -6779,7 +6778,6 @@ components:
       - performedAt
       - actionName
       - state
-      - buildInfo
       properties:
         id:
           type: string
@@ -6815,8 +6813,10 @@ components:
           type:
           - string
           - 'null'
-        buildInfo:
-          $ref: '#/components/schemas/Map_String'
+        modelInfo:
+          anyOf:
+          - $ref: '#/components/schemas/Map_String'
+          - type: 'null'
     ProcessActionState:
       title: ProcessActionState
       type: string
