+++ engine/flink/management/dev-model/src/main/scala/pl/touk/nussknacker/engine/management/sample/DevProcessConfigCreator.scala
@@ -11,6 +11,7 @@ import org.apache.flink.streaming.api.functions.sink.DiscardingSink
 import pl.touk.nussknacker.engine.api._
 import pl.touk.nussknacker.engine.api.component.{ComponentConfig, ComponentGroupName, ParameterConfig}
 import pl.touk.nussknacker.engine.api.definition._
+import pl.touk.nussknacker.engine.api.modelinfo.ModelInfo
 import pl.touk.nussknacker.engine.api.parameter.ParameterName
 import pl.touk.nussknacker.engine.api.process._
 import pl.touk.nussknacker.engine.flink.util.sink.{EmptySink, SingleValueSinkFactory}
@@ -249,11 +250,13 @@ class DevProcessConfigCreator extends ProcessConfigCreator {
   }
 
   // we generate static generation-time during ConfigCreator creation to test reload mechanisms
-  override val buildInfo: Map[String, String] = {
-    Map(
-      "process-version" -> "0.1",
-      "engine-version"  -> "0.1",
-      "generation-time" -> LocalDateTime.now().toString
+  override val modelInfo: ModelInfo = {
+    ModelInfo.fromMap(
+      Map(
+        "process-version" -> "0.1",
+        "engine-version"  -> "0.1",
+        "generation-time" -> LocalDateTime.now().toString
+      )
     )
   }
