+++ extensions-api/src/main/scala/pl/touk/nussknacker/engine/api/process/ProcessConfigCreator.scala
@@ -1,5 +1,6 @@
 package pl.touk.nussknacker.engine.api.process
 
+import pl.touk.nussknacker.engine.api.modelinfo.ModelInfo
 import pl.touk.nussknacker.engine.api.{CustomStreamTransformer, ProcessListener, Service}
 
 /**
@@ -22,8 +23,7 @@ trait ProcessConfigCreator extends Serializable {
 
   def expressionConfig(modelDependencies: ProcessObjectDependencies): ExpressionConfig
 
-  // TODO: Rename to modelInfo or similar, as it can contain any information related to model, not only build info
-  def buildInfo(): Map[String, String]
+  def modelInfo(): ModelInfo
 
   def asyncExecutionContextPreparer(
       modelDependencies: ProcessObjectDependencies
